pkg/apis/monitoring/v1: add SafeTLSConfig type
The TLSConfig structure contains unsafe configuration parameters such as
the ones accessing the filesystem. This was mitigated by using the
ArbitraryFSAccessThroughSms mechanism. However, we do not want to repeat
the same mistakes with PodMonitor and only implement safe functionalities.
Thus, the SafeTLSConfig type was created to contain all the safe TLS
configuration mechanism.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>